requirements.txt
discord.py
kwslogger
PyYAML
sqlite3